Updating c++ to take advantage of aliases.
C++11 adds the ability to call 'using' to create aliases in namespaces. JTS
by default will generate new class files for every InternalEvent and Message
for a service, even if those InternalEvents/Messages are already defined by
a parent or client service. This increases the number of files that need to
be compiled, and, for those that care, adds to the size of the library or
application.

The changes in this commit update the C++ code generation to use aliases
for InternalEvents/Messages defined by parent or client services. The
MessageSet.h and InternalEventsSet.h file will get updated to include the
parent's MessageSet.h or InternalEventsSet.h and create appropriate aliases.

These changes could probably be cleaned up a bit. They don't handle
DeclaredMessageDef or DeclaredEventDef classes, but it doesn't look like
the original code does either.