Updated Services to use std::lock_guard for C++.
The default Service implementation uses goto statements
to handle unlocking a mutex during the state transitions.
C++11 introduces the std::lock_guard<MutexType> that provides
 a convenient RAII-style mechanism for owning a mutex for
the duration of a scoped block. Using std::lock_guard, the
services can simply return when they are done processing
a state transition and the mutex will automatically be
unlocked when the std::lock_guard goes out of scope.

In order for this to work, I Needed to update the
Sconstruct generator to add '-std=c++11' to the CXXFLAGS.
Also, the service template needed to be updated.