Updated to support c++11 and legacy c++.
Updated the code to support c++11 and legacy c++. The previous
commits made an assumption that users would be switching to c++11.
However, some users may still want to use legacy c++. To account
for this, the C_PLUS_PLUS_11 codeType was added and the UI was
updated to allow selecting C++11.

The templates/service-cpp11 directory was created for the C++11
service templates. This is currently the only template that is using
C++11. The remaining templates will fallback to their pervious versions.